finished ChristmasMarketBox + added README.md
