+++ SOHChristmasMarketBox/config.json
@@ -1,15 +1,21 @@
 {
-  "id": "rathausmarkt_2023",
+  "id": "rathausmarkt_2024",
   "globals": {
     "deltaT": 1,
-    "startPoint": "2023-12-10T16:00:00",
-    "endPoint": "2023-12-10T17:00:00",
+    "startPoint": "2024-12-10T16:00:00",
+    "endPoint": "2024-12-10T18:00:00",
     "deltaTUnit": "seconds",
     "console": true
   },
-  "agents": [
-    {
-      "name": "RandomWalkMarketTraveler",
+    "agents": [
+      {
+        "name": "OptimalStepsMarketTraveler",
+        "individual": [
+          {
+            "parameter": "PreferredSpeed",
+            "value": 1.4
+        }
+      ],
       "outputs": [
         {
           "kind": "trips",
@@ -24,24 +30,7 @@
         {
           "value": true,
           "parameter": "ResultTrajectoryEnabled"
-        },
-        {
-          "parameter": "topLeftCorner",
-          "value": "9.992230367450674,53.550967416608785"
-        },
-        {
-          "parameter": "topRightCorner",
-          "value": "9.993077605872003,53.55039694865533"
-        },
-        {
-          "parameter": "bottomRightCorner",
-          "value": "9.993672555518117,53.55069672493096"
-        },
-        {
-          "parameter": "bottomLeftCorner",
-          "value": "9.992893096170604,53.55126495174483"
         }
-
       ]
     }
   ],
@@ -64,13 +53,31 @@
       "name": "MarketLayer",
       "inputs": [
         {
-          "file": "resources/weihnachtsmarkt_polygon.geojson"
+          "file": "resources/market_stalls.geojson"
+        }
+      ],
+      "individual": [
+        {
+          "parameter": "topLeftCorner",
+          "value": "9.992230367450674,53.550967416608785"
+        },
+        {
+          "parameter": "topRightCorner",
+          "value": "9.993077605872003,53.55039694865533"
+        },
+        {
+          "parameter": "bottomRightCorner",
+          "value": "9.993672555518117,53.55069672493096"
+        },
+        {
+          "parameter": "bottomLeftCorner",
+          "value": "9.992893096170604,53.55126495174483"
         }
       ]
     },
     {
       "name": "MarketTravelerSchedulerLayer",
-      "file": "resources/human_traveler_rathausmarkt_v2.csv"
+      "file": "resources/human_traveler_rathausmarkt.csv"
     }
   ]
 }
