+++ SOHChristmasMarketBox/resources/human_traveler_rathausmarkt.csv
@@ -1,8 +1,5 @@
 startTime,endTime,spawningIntervalInMinutes,spawningAmount,hasBike,hasCar,prefersBike,prefersCar,usesBikeAndsdRide,usesOwnBikeOutside,usesOwnCar,source,destination,discriminator
-15:00,20:30,3,5,0,0,0,0,0,0,0,"POLYGON((9.9935 53.5508,9.9945 53.5508,9.9945 53.5502,9.9935 53.5502,9.9935 53.5508))","POLYGON((9.9952 53.5510,9.9962 53.5510,9.9962 53.5504,9.9952 53.5504,9.9952 53.5510))",2
-15:00,20:30,3,1,0,0,0,0,0,0,0,"POLYGON((9.9935 53.5508,9.9945 53.5508,9.9945 53.5502,9.9935 53.5502,9.9935 53.5508))","POLYGON((9.9952 53.5510,9.9962 53.5510,9.9962 53.5504,9.9952 53.5504,9.9952 53.5510))",2
-15:00,20:30,3,2,0,0,0,0,0,0,0,"POLYGON((9.9930 53.5505,9.9940 53.5505,9.9940 53.5499,9.9930 53.5499,9.9930 53.5505))","POINT(9.9957 53.5506)",2
-15:00,20:30,3,18,0,0,0,0,0,0,0,"POLYGON((9.9932 53.5507,9.9942 53.5507,9.9942 53.5501,9.9932 53.5501,9.9932 53.5507))","POLYGON((9.9950 53.5508,9.9960 53.5508,9.9960 53.5502,9.9950 53.5502,9.9950 53.5508))",2
-15:00,20:30,4,15,0,0,0,0,0,0,0,"POLYGON((9.9932 53.5507,9.9942 53.5507,9.9942 53.5501,9.9932 53.5501,9.9932 53.5507))","POLYGON((9.9950 53.5508,9.9960 53.5508,9.9960 53.5502,9.9950 53.5502,9.9950 53.5508))",3
-15:00,20:30,3,18,0,0,0,0,0,0,0,"POLYGON((9.9950 53.5508,9.9960 53.5508,9.9960 53.5502,9.9950 53.5502,9.9950 53.5508))","POLYGON((9.9932 53.5507,9.9942 53.5507,9.9942 53.5501,9.9932 53.5501,9.9932 53.5507))",2
-15:00,20:30,4,14,0,0,0,0,0,0,0,"POLYGON((9.9950 53.5508,9.9960 53.5508,9.9960 53.5502,9.9950 53.5502,9.9950 53.5508))","POLYGON((9.9932 53.5507,9.9942 53.5507,9.9942 53.5501,9.9932 53.5501,9.9932 53.5507))",3
+15:00,20:30,2,10,0,0,0,0,0,0,0,"POLYGON((9.9905 53.5505,9.9920 53.5505,9.9920 53.5498,9.9905 53.5498,9.9905 53.5505))","POLYGON((9.992230367450674 53.550967416608785,9.993077605872003 53.55039694865533,9.993672555518117 53.55069672493096,9.992893096170604 53.55126495174483,9.992230367450674 53.550967416608785))",2
+15:00,20:30,2,10,0,0,0,0,0,0,0,"POLYGON((9.9900 53.5520,9.9915 53.5520,9.9915 53.5513,9.9900 53.5513,9.9900 53.5520))","POLYGON((9.992230367450674 53.550967416608785,9.993077605872003 53.55039694865533,9.993672555518117 53.55069672493096,9.992893096170604 53.55126495174483,9.992230367450674 53.550967416608785))",3
+15:00,20:30,2,10,0,0,0,0,0,0,0,"POLYGON((9.9940 53.5495,9.9955 53.5495,9.9955 53.5488,9.9940 53.5488,9.9940 53.5495))","POLYGON((9.992230367450674 53.550967416608785,9.993077605872003 53.55039694865533,9.993672555518117 53.55069672493096,9.992893096170604 53.55126495174483,9.992230367450674 53.550967416608785))",4
+15:00,20:30,2,10,0,0,0,0,0,0,0,"POLYGON((9.9950 53.5515,9.9965 53.5515,9.9965 53.5508,9.9950 53.5508,9.9950 53.5515))","POLYGON((9.992230367450674 53.550967416608785,9.993077605872003 53.55039694865533,9.993672555518117 53.55069672493096,9.992893096170604 53.55126495174483,9.992230367450674 53.550967416608785))",5
