--- SOHChristmasMarketBox/resources/human_traveler_rathausmarkt_v2.csv
@@ -1,4 +1,0 @@
-startTime,endTime,spawningIntervalInMinutes,spawningAmount,hasBike,hasCar,prefersBike,prefersCar,usesBikeAndsdRide,usesOwnBikeOutside,usesOwnCar,source,destination,discriminator
-15:00,20:30,3,4,0,0,0,0,0,0,0,"POLYGON((9.9915 53.5505,9.9930 53.5505,9.9930 53.5498,9.9915 53.5498,9.9915 53.5505))","POLYGON((9.9908 53.5545,9.9940 53.5545,9.9940 53.5535,9.9908 53.5535,9.9908 53.5545))",2
-15:00,20:30,3,4,0,0,0,0,0,0,0,"POLYGON((9.9908 53.5545,9.9940 53.5545,9.9940 53.5535,9.9908 53.5535,9.9908 53.5545))","POLYGON((9.9915 53.5505,9.9930 53.5505,9.9930 53.5498,9.9915 53.5498,9.9915 53.5505))",2
-15:00,20:30,4,2,0,0,0,0,0,0,0,"POLYGON((9.9935 53.5495,9.9955 53.5495,9.9955 53.5485,9.9935 53.5485,9.9935 53.5495))","POLYGON((9.9915 53.5505,9.9930 53.5505,9.9930 53.5498,9.9915 53.5498,9.9915 53.5505))",3
