--- SOHChristmasMarketBox/resources/weihnachtsmarkt_polygon.geojson
@@ -1,1 +1,0 @@
-{"type":"FeatureCollection","features":[{"type":"Feature","properties":{},"geometry":{"coordinates":[[[9.992191167747592,53.550982371138815],[9.993046531947641,53.550402920201776],[9.993714133274835,53.55070039333174],[9.992910925427907,53.55127674157012],[9.992191167747592,53.550982371138815]]],"type":"Polygon"}}]}
