--- SOHChristmasMarketBox/resources/weihnachtsmarkt_staende.csv
@@ -1,6 +1,0 @@
-name,type,longitude,latitude
-Glühwein-Paradies,Drinks,9.99185,53.55153
-Bratwurst-König,Food,9.99248,53.55182
-Heisse Mandeln,Sweets,9.99215,53.55219
-Erzgebirgskunst,Crafts,9.99287,53.55204
-Lederwaren,Gifts,9.99266,53.55161
