+++ SOHModel/ChristmasMarket/Entities/MarketStallType.cs
@@ -1,0 +1,8 @@
+namespace SOHModel.ChristmasMarket;
+
+public enum MarketStallType
+{
+    Glühwein,
+    Gastronomie,
+    Verkaufsstand
+}
