--- SOHModel/ChristmasMarket/IMarketTravelerLayer.cs
@@ -1,8 +1,0 @@
-using SOHModel.Multimodal.Multimodal;
-
-namespace SOHModel.Multimodal.Model;
-
-public interface IMarketTravelerLayer : IMultimodalLayer
-{
-    
-}
