--- SOHModel/ChristmasMarket/MarketStallType.cs
@@ -1,9 +1,0 @@
-namespace SOHModel.Domain.Model;
-
-public enum MarketStallType
-{
-    Food,
-    Drinks,
-    Crafts,
-    Gifts
-}
