--- SOHModel/ChristmasMarket/SocialForcesMarketTraveler.cs
@@ -1,10 +1,0 @@
-using SOHModel.Multimodal.Model;
-
-namespace SOHModel.ChristmasMarket;
-
-public class SocialForcesMarketTraveler : MarketTraveler
-{
-    protected override void SimulateFreeMovement()
-    {
-    }
-}
